seralize launch process test
