+++ test/dotnetup.Tests/DnupCollections.cs
@@ -62,3 +62,14 @@ public class DotnetupEnvironmentMutationTests
 public class DotnetupTelemetryStateMutationTests
 {
 }
+
+/// <summary>
+/// Serialized collection for tests that launch the real dotnetup executable as a child
+/// process with redirected stdin/stdout. These tests are timing-sensitive (process startup,
+/// PATH lookup, pipe drain) and have shown intermittent flakes when run concurrently with
+/// other CPU-heavy or filesystem-heavy tests. Serializing them removes that contention.
+/// </summary>
+[CollectionDefinition("DotnetupProcessLaunchTests", DisableParallelization = true)]
+public class DotnetupProcessLaunchTests
+{
+}
