+++ test/dotnetup.Tests/DotnetCommandTests.cs
@@ -283,77 +283,6 @@ public class DotnetCommandTests
         }
     }
 
-    // ── Stdin forwarding (interactivity) tests ───────────────────────────
-
-    /// <summary>
-    /// Verifies that interactive stdin data is properly forwarded from the
-    /// parent process through dotnetup to the child dotnet process.
-    /// This is critical for commands like <c>dotnet nuget push --interactive</c>
-    /// or <c>dotnet new</c> that prompt for user input.
-    ///
-    /// The test launches dotnetup as a real process, redirects its stdin,
-    /// writes test data, and verifies the child process echoes it back
-    /// through stdout — proving the full stdin forwarding pipeline works.
-    /// </summary>
-    [Fact]
-    public void DotnetCommand_ForwardsStdinToChildProcess()
-    {
-        var tempDir = Directory.CreateTempSubdirectory("dotnetup-stdin-test");
-        try
-        {
-            CreateStdinEchoFakeDotnet(tempDir.FullName);
-
-            string dotnetupPath = DotnetupTestUtilities.GetDotnetupExecutablePath();
-            string testInput = "HelloInteractive_" + Guid.NewGuid().ToString("N")[..8];
-
-            using var process = new Process();
-            process.StartInfo.FileName = dotnetupPath;
-            process.StartInfo.UseShellExecute = false;
-            process.StartInfo.CreateNoWindow = true;
-            process.StartInfo.RedirectStandardInput = true;
-            process.StartInfo.RedirectStandardOutput = true;
-            process.StartInfo.RedirectStandardError = true;
-
-            // Prepend our temp dir to PATH so dotnetup resolves our fake dotnet
-            var currentPath = Environment.GetEnvironmentVariable("PATH") ?? string.Empty;
-            process.StartInfo.Environment["PATH"] = tempDir.FullName + Path.PathSeparator + currentPath;
-            process.StartInfo.Environment["DOTNET_NOLOGO"] = "1";
-            process.StartInfo.Environment["NO_COLOR"] = "1";
-
-            if (OperatingSystem.IsWindows())
-            {
-                // Fake dotnet.exe is cmd.exe; forward /c "findstr /r ." to echo stdin.
-                // findstr reads from stdin and echoes lines matching regex "." (non-empty).
-                process.StartInfo.Arguments = ArgumentEscaper.EscapeAndConcatenateArgArrayForProcessStart(
-                    ["dotnet", "/c", "findstr /r ."]);
-            }
-            else
-            {
-                // Fake dotnet is a shell script that cats stdin to stdout.
-                process.StartInfo.Arguments = "dotnet";
-            }
-
-            process.Start();
-
-            // Write test data to dotnetup's stdin; the child process inherits this
-            // stdin handle because DotnetCommand uses RedirectStandardInput = false.
-            process.StandardInput.WriteLine(testInput);
-            process.StandardInput.Close();
-
-            string output = process.StandardOutput.ReadToEnd();
-            string stderr = process.StandardError.ReadToEnd();
-            process.WaitForExit();
-
-            // The child process should have received and echoed our input
-            output.Should().Contain(testInput,
-                because: "stdin data should be forwarded to the child dotnet process for interactive commands");
-        }
-        finally
-        {
-            try { tempDir.Delete(recursive: true); } catch { /* cleanup best-effort */ }
-        }
-    }
-
     // ── Helpers ───────────────────────────────────────────────────────────
 
     /// <summary>
@@ -375,7 +304,7 @@ public class DotnetCommandTests
     /// On Windows, copies cmd.exe as dotnet.exe (caller passes /c "findstr /r ." to echo stdin).
     /// On Unix, creates a shell script that cats stdin to stdout.
     /// </summary>
-    private static void CreateStdinEchoFakeDotnet(string directory)
+    internal static void CreateStdinEchoFakeDotnet(string directory)
     {
         if (OperatingSystem.IsWindows())
         {
